Cloud_Functions_Option_01/requirements.txt
pandas
numpy
requests
google-cloud-bigquery
google-cloud-storage
google-cloud-secret-manager
python-dotenv
